# perf: Improve performance especially of validate and clone