--- a/lib/document.js
+++ b/lib/document.js
@@ -1135,7 +1135,7 @@ Document.prototype.$set = function $set(path, val, type, options) {
       // them to ensure we keep the user's key order.
       if (type === true &&
           !prefix &&
-          path[key] != null &&
+          valForKey != null &&
           pathtype === 'nested' &&
           this._doc[key] != null) {
         delete this._doc[key];
@@ -1313,7 +1313,7 @@ Document.prototype.$set = function $set(path, val, type, options) {
   let curPath = '';
   for (i = 0; i < parts.length - 1; ++i) {
     cur = cur[parts[i]];
-    curPath += (curPath.length > 0 ? '.' : '') + parts[i];
+    curPath += (curPath.length !== 0 ? '.' : '') + parts[i];
     if (!cur) {
       this.$set(curPath, {});
       // Hack re: gh-5800. If nested field is not selected, it probably exists
@@ -1333,10 +1333,11 @@ Document.prototype.$set = function $set(path, val, type, options) {
   // When using the $set operator the path to the field must already exist.
   // Else mongodb throws: "LEFT_SUBFIELD only supports Object"
 
-  if (parts.length <= 1) {
+  if (parts.length < 2) {
     pathToMark = path;
   } else {
-    for (i = 0; i < parts.length; ++i) {
+    const len = parts.length;
+    for (i = 0; i < len; ++i) {
       const subpath = parts.slice(0, i + 1).join('.');
       if (this.$get(subpath, null, { getters: false }) === null) {
         pathToMark = subpath;
@@ -1361,7 +1362,7 @@ Document.prototype.$set = function $set(path, val, type, options) {
     _markValidSubpaths(this, path);
   }
 
-  if (schema.$isSingleNested && val != null && merge) {
+  if (val != null && merge && schema.$isSingleNested) {
     if (val instanceof Document) {
       val = val.toObject({ virtuals: false, transform: false });
     }
@@ -1431,9 +1432,10 @@ Document.prototype.$set = function $set(path, val, type, options) {
       val = schema.applySetters(val, this, false, priorVal);
     }
 
-    if (schema.$isMongooseDocumentArray &&
-        Array.isArray(val) &&
-        val.length > 0 &&
+    if (Array.isArray(val) &&
+        !Array.isArray(schema) &&
+        schema.$isMongooseDocumentArray &&
+        val.length !== 0 &&
         val[0] != null &&
         val[0].$__ != null &&
         val[0].$__.populated != null) {
@@ -1459,7 +1461,7 @@ Document.prototype.$set = function $set(path, val, type, options) {
       delete this.$__.populated[path];
     }
 
-    if (schema.$isSingleNested && val != null) {
+    if (val != null && schema.$isSingleNested) {
       _checkImmutableSubpaths(val, schema, priorVal);
     }
 
@@ -1652,11 +1654,11 @@ Document.prototype.$__set = function(pathToMark, path, options, constructing, pa
 
     // handle directly setting arrays (gh-1126)
     MongooseArray || (MongooseArray = require('./types/array'));
-    if (val && val.isMongooseArray) {
+    if (val && utils.isMongooseArray(val)) {
       val._registerAtomic('$set', val);
 
       // Update embedded document parent references (gh-5189)
-      if (val.isMongooseDocumentArray) {
+      if (utils.isMongooseDocumentArray(val)) {
         val.forEach(function(item) {
           item && item.__parentArray && (item.__parentArray = val);
         });
@@ -1670,10 +1672,10 @@ Document.prototype.$__set = function(pathToMark, path, options, constructing, pa
         }
       });
     }
-  } else if (Array.isArray(val) && val.isMongooseArray && Array.isArray(priorVal) && priorVal.isMongooseArray) {
+  } else if (Array.isArray(val) && Array.isArray(priorVal) && utils.isMongooseArray(val) && utils.isMongooseArray(priorVal)) {
     val[arrayAtomicsSymbol] = priorVal[arrayAtomicsSymbol];
     val[arrayAtomicsBackupSymbol] = priorVal[arrayAtomicsBackupSymbol];
-    if (val.isMongooseDocumentArray) {
+    if (utils.isMongooseDocumentArray(val)) {
       val.forEach(doc => { doc.isNew = false; });
     }
   }
@@ -1702,7 +1704,7 @@ Document.prototype.$__set = function(pathToMark, path, options, constructing, pa
         obj = obj[parts[i]];
       } else if (obj[parts[i]] && obj[parts[i]] instanceof Embedded) {
         obj = obj[parts[i]];
-      } else if (obj[parts[i]] && obj[parts[i]].$isSingleNested) {
+      } else if (obj[parts[i]] && !Array.isArray(obj[parts[i]]) && obj[parts[i]].$isSingleNested) {
         obj = obj[parts[i]];
       } else if (obj[parts[i]] && Array.isArray(obj[parts[i]])) {
         obj = obj[parts[i]];
@@ -1963,7 +1965,7 @@ Document.prototype.$isEmpty = function(path) {
     transform: false
   };
 
-  if (arguments.length > 0) {
+  if (arguments.length !== 0) {
     const v = this.$get(path);
     if (v == null) {
       return true;
@@ -2006,51 +2008,60 @@ function _isEmpty(v) {
 
 Document.prototype.modifiedPaths = function(options) {
   options = options || {};
-  const directModifiedPaths = Object.keys(this.$__.activePaths.states.modify);
-  const _this = this;
-  return directModifiedPaths.reduce(function(list, path) {
-    const parts = path.split('.');
-    list = list.concat(parts.reduce(function(chains, part, i) {
-      return chains.concat(parts.slice(0, i).concat(part).join('.'));
-    }, []).filter(function(chain) {
-      return (list.indexOf(chain) === -1);
-    }));
 
-    if (!options.includeChildren) {
-      return list;
+  const directModifiedPaths = Object.keys(this.$__.activePaths.states.modify);
+  const result = new Set();
+
+  let i = 0;
+  let j = 0;
+  const len = directModifiedPaths.length;
+
+  for (i = 0; i < len; ++i) {
+    const path = directModifiedPaths[i];
+    const parts = parentPaths(path);
+    const pLen = parts.length;
+
+    for (j = 0; j < pLen; ++j) {
+      result.add(parts[j]);
     }
 
-    let cur = _this.$get(path);
-    if (cur != null && typeof cur === 'object') {
+    if (!options.includeChildren) {
+      continue;
+    }
+
+    let ii = 0;
+    let cur = this.$get(path);
+    if (typeof cur === 'object' && cur !== null) {
       if (cur._doc) {
         cur = cur._doc;
       }
+      const len = cur.length;
       if (Array.isArray(cur)) {
-        const len = cur.length;
-        for (let i = 0; i < len; ++i) {
-          if (list.indexOf(path + '.' + i) === -1) {
-            list.push(path + '.' + i);
-            if (cur[i] != null && cur[i].$__) {
-              const modified = cur[i].modifiedPaths();
-              for (const childPath of modified) {
-                list.push(path + '.' + i + '.' + childPath);
+        for (ii = 0; ii < len; ++ii) {
+          const subPath = path + '.' + ii;
+          if (!result.has(subPath)) {
+            result.add(subPath);
+            if (cur[ii] != null && cur[ii].$__) {
+              const modified = cur[ii].modifiedPaths();
+              let iii = 0;
+              const iiiLen = modified.length;
+              for (iii = 0; iii < iiiLen; ++iii) {
+                result.add(subPath + '.' + modified[iii]);
               }
             }
           }
         }
       } else {
-        Object.keys(cur).
-          filter(function(key) {
-            return list.indexOf(path + '.' + key) === -1;
-          }).
-          forEach(function(key) {
-            list.push(path + '.' + key);
-          });
+        const keys = Object.keys(cur);
+        let ii = 0;
+        const len = keys.length;
+        for (ii = 0; ii < len; ++ii) {
+          result.add(path + '.' + keys[ii]);
+        }
       }
     }
-
-    return list;
-  }, []);
+  }
+  return Array.from(result);
 };
 
 Document.prototype[documentModifiedPaths] = Document.prototype.modifiedPaths;
@@ -2448,7 +2459,12 @@ Document.prototype.$validate = Document.prototype.validate;
  */
 
 function _evaluateRequiredFunctions(doc) {
-  Object.keys(doc.$__.activePaths.states.require).forEach(path => {
+  const requiredFields = Object.keys(doc.$__.activePaths.states.require);
+  let i = 0;
+  const len = requiredFields.length;
+  for (i = 0; i < len; ++i) {
+    const path = requiredFields[i];
+
     const p = doc.$__schema.path(path);
 
     if (p != null && typeof p.originalRequiredValue === 'function') {
@@ -2459,7 +2475,7 @@ function _evaluateRequiredFunctions(doc) {
         doc.invalidate(path, err);
       }
     }
-  });
+  }
 }
 
 /*!
@@ -2520,7 +2536,7 @@ function _getPathsToValidate(doc) {
         // To avoid potential performance issues, skip doc arrays whose children
         // are not required. `getPositionalPathType()` may be slow, so avoid
         // it unless we have a case of #6364
-        (_pathType.$isMongooseDocumentArray && !get(_pathType, 'schemaOptions.required'))) {
+        (!Array.isArray(_pathType) && _pathType.$isMongooseDocumentArray && !get(_pathType, 'schemaOptions.required'))) {
       continue;
     }
 
@@ -3112,7 +3128,7 @@ Document.prototype.$__reset = function reset() {
       return _this.$__getValue(i);
     })
     .filter(function(val) {
-      return val && val instanceof Array && val.isMongooseDocumentArray && val.length;
+      return val && val instanceof Array && utils.isMongooseDocumentArray(val) && val.length;
     })
     .forEach(function(array) {
       let i = array.length;
@@ -3135,7 +3151,7 @@ Document.prototype.$__reset = function reset() {
       return _this.$__getValue(i);
     }).
     filter(function(val) {
-      return val && val.$isSingleNested;
+      return val && !Array.isArray(val) && val.$isSingleNested;
     }).
     forEach(function(doc) {
       doc.$__reset();
@@ -3315,7 +3331,7 @@ Document.prototype.$__getArrayPathsToValidate = function() {
       return this.$__getValue(i);
     }.bind(this))
     .filter(function(val) {
-      return val && val instanceof Array && val.isMongooseDocumentArray && val.length;
+      return val && val instanceof Array && utils.isMongooseDocumentArray(val) && val.length;
     }).reduce(function(seed, array) {
       return seed.concat(array);
     }, [])
@@ -3357,12 +3373,12 @@ Document.prototype.$getAllSubdocs = function() {
       seed = Array.from(val.keys()).reduce(function(seed, path) {
         return docReducer(val.get(path), seed, null);
       }, seed);
-    } else if (val && val.$isSingleNested) {
+    } else if (val && !Array.isArray(val) && val.$isSingleNested) {
       seed = Object.keys(val._doc).reduce(function(seed, path) {
         return docReducer(val._doc, seed, path);
       }, seed);
       seed.push(val);
-    } else if (val && val.isMongooseDocumentArray) {
+    } else if (val && utils.isMongooseDocumentArray(val)) {
       val.forEach(function _docReduce(doc) {
         if (!doc || !doc._doc) {
           return;
@@ -4139,7 +4155,7 @@ Document.prototype.populate = function populate() {
   const args = [...arguments];
   let fn;
 
-  if (args.length > 0) {
+  if (args.length !== 0) {
     if (typeof args[args.length - 1] === 'function') {
       fn = args.pop();
     }