--- a/lib/document.js
+++ b/lib/document.js
@@ -1333,7 +1333,7 @@ Document.prototype.$set = function $set(path, val, type, options) {
   // When using the $set operator the path to the field must already exist.
   // Else mongodb throws: "LEFT_SUBFIELD only supports Object"
 
-  if (parts.length < 2) {
+  if (parts.length <= 1) {
     pathToMark = path;
   } else {
     const len = parts.length;