--- a/lib/model.js
+++ b/lib/model.js
@@ -740,7 +740,7 @@ Model.prototype.$__delta = function() {
       operand(this, where, delta, data, 1, '$unset');
     } else if (value === null) {
       operand(this, where, delta, data, null);
-    } else if (value.isMongooseArray && value.$path() && value[arrayAtomicsSymbol]) {
+    } else if (utils.isMongooseArray(value) && value.$path() && value[arrayAtomicsSymbol]) {
       // arrays and other custom types (support plugins etc)
       handleAtomics(this, where, delta, data, value);
     } else if (value[MongooseBuffer.pathSymbol] && Buffer.isBuffer(value)) {
@@ -792,7 +792,7 @@ function checkDivergentArray(doc, path, array) {
     }
   }
 
-  if (!(pop && array && array.isMongooseArray)) return;
+  if (!(pop && utils.isMongooseArray(array))) return;
 
   // If the array was populated using options that prevented all
   // documents from being returned (match, skip, limit) or they
@@ -3374,7 +3374,7 @@ Model.$__insertMany = function(arr, options, callback) {
       return doc != null;
     });
     // Quickly escape while there aren't any valid docAttributes
-    if (docAttributes.length < 1) {
+    if (docAttributes.length === 0) {
       if (rawResult) {
         const res = {
           mongoose: {
@@ -4524,7 +4524,7 @@ function populate(model, docs, options, callback) {
   }
   if (!hasOne) {
     // If models but no docs, skip further deep populate.
-    if (modelsMap.length > 0) {
+    if (modelsMap.length !== 0) {
       return callback();
     }
     // If no models to populate but we have a nested populate,