--- a/lib/plugins/trackTransaction.js
+++ b/lib/plugins/trackTransaction.js
@@ -47,10 +47,10 @@ function _getAtomics(doc, previous) {
     const val = doc.$__getValue(path);
     if (val != null &&
         val instanceof Array &&
-        val.isMongooseDocumentArray &&
+        utils.isMongooseDocumentArray(val) &&
         val.length &&
         val[arrayAtomicsSymbol] != null &&
-        Object.keys(val[arrayAtomicsSymbol]).length > 0) {
+        Object.keys(val[arrayAtomicsSymbol]).length !== 0) {
       const existing = previous.get(path) || {};
       pathToAtomics.set(path, mergeAtomics(existing, val[arrayAtomicsSymbol]));
     }
@@ -61,7 +61,7 @@ function _getAtomics(doc, previous) {
     const path = dirt.path;
 
     const val = dirt.value;
-    if (val != null && val[arrayAtomicsSymbol] != null && Object.keys(val[arrayAtomicsSymbol]).length > 0) {
+    if (val != null && val[arrayAtomicsSymbol] != null && Object.keys(val[arrayAtomicsSymbol]).length !== 0) {
       const existing = previous.get(path) || {};
       pathToAtomics.set(path, mergeAtomics(existing, val[arrayAtomicsSymbol]));
     }