--- a/lib/queryhelpers.js
+++ b/lib/queryhelpers.js
@@ -212,7 +212,7 @@ exports.applyPaths = function applyPaths(fields, schema) {
 
       let addedPath = analyzePath(path, type);
       // arrays
-      if (addedPath == null && type.$isMongooseArray && !type.$isMongooseDocumentArray) {
+      if (addedPath == null && !Array.isArray(type) && type.$isMongooseArray && !type.$isMongooseDocumentArray) {
         addedPath = analyzePath(path, type.caster);
       }
       if (addedPath != null) {