--- a/lib/schema/array.js
+++ b/lib/schema/array.js
@@ -112,14 +112,12 @@ function SchemaArray(key, cast, options, schemaOptions) {
 
   if (!('defaultValue' in this) || this.defaultValue !== void 0) {
     const defaultFn = function() {
-      let arr = [];
-      if (fn) {
-        arr = defaultArr.call(this);
-      } else if (defaultArr != null) {
-        arr = arr.concat(defaultArr);
-      }
       // Leave it up to `cast()` to convert the array
-      return arr;
+      return fn
+      ? defaultArr.call(this)
+      : defaultArr != null
+        ? [].concat(defaultArr)
+        : [];
     };
     defaultFn.$runBeforeSetters = !fn;
     this.default(defaultFn);
@@ -275,7 +273,7 @@ SchemaArray.prototype.applyGetters = function(value, scope) {
 
   const ret = SchemaType.prototype.applyGetters.call(this, value, scope);
   if (Array.isArray(ret)) {
-    const rawValue = ret.isMongooseArrayProxy ? ret.__array : ret;
+    const rawValue = utils.isMongooseArray(ret) ? ret.__array : ret;
     const len = rawValue.length;
     for (let i = 0; i < len; ++i) {
       rawValue[i] = this.caster.applyGetters(rawValue[i], scope);
@@ -299,7 +297,7 @@ SchemaArray.prototype._applySetters = function(value, scope, init, priorVal) {
     }
 
     // No need to wrap empty arrays
-    if (value != null && value.length > 0) {
+    if (value != null && value.length !== 0) {
       const valueDepth = arrayDepth(value);
       if (valueDepth.min === valueDepth.max && valueDepth.max < depth && valueDepth.containsNonArrayItem) {
         for (let i = valueDepth.max; i < depth; ++i) {
@@ -357,7 +355,7 @@ SchemaArray.prototype.cast = function(value, doc, init, prev, options) {
 
     options = options || emptyOpts;
 
-    let rawValue = value.isMongooseArrayProxy ? value.__array : value;
+    let rawValue = utils.isMongooseArray(value) ? value.__array : value;
     value = MongooseArray(rawValue, options.path || this._arrayPath || this.path, doc, this);
     rawValue = value.__array;