--- a/lib/types/array/isMongooseArray.js
+++ b/lib/types/array/isMongooseArray.js
@@ -1,0 +1,3 @@
+exports.isMongooseArray = function (mongooseArray) {
+    return Array.isArray(mongooseArray) && mongooseArray.isMongooseArray;
+}