This PR improves the performance of especially validate and clone. 

before: 

```
MONGOOSE_DEV=1 node validate
invalid x 18,325 ops/sec ±1.40% (89 runs sampled)
valid x 46,240 ops/sec ±1.30% (91 runs sampled)

MONGOOSE_DEV=1 node clone
Simple x 124,034 ops/sec ±1.41% (91 runs sampled)
AllSchema x 17,409 ops/sec ±0.52% (94 runs sampled)

MONGOOSE_DEV=1 node create
CheckItem x 113,922 ops/sec ±0.34% (94 runs sampled)
Board x 112 ops/sec ±1.26% (74 runs sampled)

node populate
1436 completed queries on mongoose version 6.1.8
```

after:
```
MONGOOSE_DEV=1 node validate
invalid x 24,576 ops/sec ±0.89% (88 runs sampled)
valid x 119,331 ops/sec ±0.12% (95 runs sampled)

MONGOOSE_DEV=1 node clone
Simple x 148,691 ops/sec ±1.36% (87 runs sampled)
AllSchema x 22,479 ops/sec ±0.91% (95 runs sampled)

MONGOOSE_DEV=1 node create
CheckItem x 112,564 ops/sec ±0.26% (93 runs sampled)
Board x 122 ops/sec ±0.45% (77 runs sampled)

node populate
1528 completed queries on mongoose version 6.1.8
1528 completed queries on mongoose version 6.1.8

```