@Uzlopak 2022-01-30
I think the biggest performance bottleneck is the ES6 Proxy in DocumentArray and MongooseArray. Playing currently aroung with using WeakSets to see if the isMongooseArray and isMongooseDocumentArray can be boosted or not. 

Also trying to figure out if we can reduce the roundtrips through the proxy layer by using forEach etc. when iterating through the elements of those special arrays.